Reorganize and clarify permission classes
