+++ edx_rest_framework_extensions/permissions/__init__.py
@@ -1,0 +1,7 @@
+# Support backward compatibility of imports prior to creating a permissions directory
+from .redirect import LoginRedirectIfUnauthenticated
+from .basic import (
+  IsSuperuser,
+  IsStaff,
+  IsUserInUrl
+)
